Move `index` to indexing.py module
